Fix atomics ops download fail during windows build (#31)
This issue adresses the following reported issue:
https://github.com/AcademySoftwareFoundation/OpenRV/issues/25

**Problem:**
The download of the libatomic_ops is now failing the hash verification
causing a build break on Windows.

**Cause:**
The windows build of RV depends on a recent version of libatomic_ops so
that --disable-gpl can be used. To do so we were downloading the head of
the repo while verifying it against a fixed hash. This hash is not
longer valid since a recent commit in the libatomic_ops repo thus the
error.

**Solution:**
Now downloading the libatomic_ops archive from a specific commit
(https://github.com/ivmai/libatomic_ops/commit/044573903530c4a8e8318e20a830d4a0531b2035)
instead of the HEAD of the repo.

Also in this commit:
Adding two .gitignore files to prevent generated files from being added
inadvertently to the repo.

Signed-off-by: Bernard Laberge <bernard.laberge@autodesk.com>