Fix switch node EDL timings (#50)
This is bringing back a fix I did earlier, but after the branch off for
OpenRV, so it wasn't part of the initial OpenRV code. The issue is that
when we have a switch node. Here's my original PR description:

This branch fixes the timing when a Sequence node without autoEDL is
connected to a SourceNode via a SwitchNode. Normally, the SwitchNode is
computing an offset for autoEDL mode so that its frames are always
starting at 1 (so it can basically unwrap the autoEDL behaviour to find
the real frames). With autoEDL though, raw frame numbers are input so we
do not want this. The simplest fix I can think of is to pass the autoEDL
mode in the context when we metaEvaluate the graph when computing the
ranges. This way if we are connected to a Sequence node with autoEDL
disabled, we can ignore the offset.

Signed-off-by: Roger Nelson <roger.nelson@autodesk.com>