SG-29775 Fix RvSync broken in Open RV (#51)
SG-29775: Fix RvSync sessions broken due to min RV version in Open RV

This commit fixes the following issue:
https://github.com/AcademySoftwareFoundation/OpenRV/issues/44

#### Problem:
The RV Sync package has a check to validate remotely connected contacts
against a minimum version of RV (eg. RV 2021.0.3). This check fails on
Open RV because the Open RV version has been reset to 1.0.0.
Note that the goal here is to find a solution where participants of an
RV Sync session can be a mix of Open RV and Commercial RV users.

#### Solution:
Check against the internal version of the Rv Sync package instead of a
minimum RV version like it was done before.

#### Rationale:
Even though the Commercial version of RV will eventually be built off
Open RV, it is assumed that their versioning will diverge. For RV
packages, a new minimal Open RV version requirement (named openrv) was
added to the PACKAGE file to complement the minimal Commercial RV
version requirement (named rv). This way, even though both Commercial
and Open RV will have different versions, the goal is for RV packages to
be the same for both. Therefore it is the RV Sync package version that
should be checked against each other to see if they are compatible or
not.

In this commit, the currently minimum version of RV specified in the
Sync.mu code was transferred to the Sync' PACKAGE file. The internal
version of the sync package was bumped and the checkRVVersion was
replaced with checkSyncVersion().

Note that I tried to find a solution that would be forward compatible so
that an Open RV sync session could be joined by Commercial RV (old
version <= 2022.3.1) users. However I was not able to because the
Commercial RV legacy implementation remoteSendEvent the remote RV to
call it back with the RV version so no sync.mu code gets executed on the
remote RV (which I could have intercepted to implement some forward
compatibility).

```
remoteSendEvent("remote-eval", "*", """
            require system; remoteSendEvent("remote-eval", "*", 
                "require sync; sync.checkRVVersion(\\"%s\\");" % system.getenv("TWK_APP_VERSION"));
            """);
```

However, the upcoming RV Commercial 2023.0 will contain the exact same
sync.mu as the one in Open RV so they ARE going to be compatible.
Until RV Commercial 2023.0 is released, as a work around, one could
simply copy the new sync.mu over to the Commercial RV.

Signed-off-by: Bernard Laberge <bernard.laberge@autodesk.com>