SG-29809: Fix unable to write RV settings on Windows (#55)
SG-29809: Fix unable to write RV settings on Windows

1. Description of changes
This commit fixes the following Open RV reported issue on Windows:
https://github.com/AcademySoftwareFoundation/OpenRV/issues/53

#### Problem:
RV settings (QSettings) default location on Windows is obsolete
([home]/Application Data)
Qt automatically falls back to the default [home]/AppData location but
in some environment/users, an error is reported. I was not able to
reproduce this error and even the user who reported it said that it was
giving an error on his laptop but not on his workstation.

#### Solution:
Now using
QStandardPaths::writableLocation(QStandardPaths::AppDataLocation) which
maps to [home]AppData/Roaming on Windows and validated that the
OpenRV.ini file is correctly saved to the same location as before
(AppData/Roaming).

Note: A roaming profile has all the functionality of a local profile but
can be transferred from one PC to another. Essentially, the user’s
profile and files are downloaded to any computer that they log onto.
When the user logs off, the changes in a roaming profile are
synchronized with the server copy of the profile. This cycle repeats
itself when the user logs on to their profile.

2. Automation
Not covered by automated tests

3. Any tips for QA for testing?
Please validate the the RV.ini file is still stored at the same location
on Windows:
For Open RV:
[home]/AppData/Roaming/Autodesk/OpenRV.ini
For Commercial RV:
[home]/AppData/Roaming/TweakSoftware/RV.ini
Thanks

Signed-off-by: Bernard Laberge <bernard.laberge@autodesk.com>