Allow building OpenRV as a subproject (#58)
This branch allows building OpenRV as a subproject. 

The problem with using the `CMAKE_SOURCE_DIR` variable is that if you
add OpenRV. as a subproject the `CMAKE_SOURCE_DIR` is the source dir of
the parent project. By using `PROJECT_SOURCE_DIR` we'll ensure all the
paths are relative to the OpenRV root, whether or not it is a
subproject.

Because the versioning was being added in one of the CMakeFiles in the
module path, I had to move that into the root CMakeFile where the
`project` call is done.

---------

Signed-off-by: Roger Nelson <roger.nelson@autodesk.com>