Fix typo in Linux documentation (#83)
<!--
Thanks for your contribution! Please read this comment in its entirety.
It's quite important.
When a contributor merges the pull request, the title and the
description will be used to build the merge commit!

### Pull Request TITLE

It should be in the following format:

[ 12345: Summary of the changes made ] Where 12345 is the corresponding
Github Issue

OR

[ Summary of the changes made ] If it's solving something trivial, like
fixing a typo.
-->

### Linked issues
<!--
Link the Issue(s) this Pull Request is related to.

Each PR should link to at least one issue, in the form:

Use one line for each Issue. This allows auto-closing the related issue
when the fix is merged.

Fixes #12345
Fixes #54345
-->

https://github.com/AcademySoftwareFoundation/OpenRV/discussions/60

### Summarize your change.

Fixes an issue repoted by
[@rubikstriangle](https://github.com/rubikstriangle) and resolved by
[tjjackson](https://github.com/tjjackson) where the ncurse-devel is not
properly defined.

The right package name is ncurses-devel, as defined
[here](https://centos.pkgs.org/7/centos-x86_64/ncurses-devel-5.9-14.20130511.el7_4.x86_64.rpm.html)

### Describe the reason for the change.

Typo.

### Describe what you have tested and on which operating system.

Linux

### Add a list of changes, and note any that might need special
attention during the review.

N/A

### If possible, provide screenshots.

N/A

Signed-off-by: Kerby Geffrard <kerby.geffrard@autodesk.com>