Update CODEOWNERS (#108)
<!--
Thanks for your contribution! Please read this comment in its entirety.
It's quite important.
When a contributor merges the pull request, the title and the
description will be used to build the merge commit!

### Pull Request TITLE

It should be in the following format:

[ 12345: Summary of the changes made ] Where 12345 is the corresponding
Github Issue

OR

[ Summary of the changes made ] If it's solving something trivial, like
fixing a typo.
-->

### Summarize your change.

Update the CODEOWNER file so the main contributors get automatically
assigned to the pull requests. Other contributors can still approve pull
requests if granted write permissions to the repository.

### Describe the reason for the change.

The previous CODEOWNER file needed to be fixed as expected. (I was under
the impression that groups were a thing with CODEOWNER files...)

Signed-off-by: Kerby Geffrard <kerby.geffrard@gmail.com>