Readthedocs integration (#109)
### Linked issues

_none_

### Summarize your change.

Update the documentation to allow publication on
[readthedocs](https://aswf-openrv.readthedocs.io/en/latest/).

### Describe the reason for the change.

Simplify access to the Open RV documentation through readthedocs. This
should improve legibility and searchability.

### Describe what you have tested and on which operating system.

Couldn't test live on the read the docs system since it integrates
directly with the github repo.
Did use sphinx to build and render the docs locally. While the pipeline
is obviously different, the rendering is similar. Local rendering was
successful.

### Add a list of changes, and note any that might need special
attention during the review.

Files required by readthedocs:
- .readthedocs.yml
- docs/conf.py
- docs/requirements.txt

Files modified to create the documentation structure on readthedocs, aka
the TOC:
- docs/index.md

The deleted files were previously used to provide a sort of toc for
navigating within github, but are no longer required for readthedocs.

---------

Signed-off-by: Martin Chesnay <Martin.Chesnay@autodesk.com>