100: Fixing MacOS build with CMake 3.26 (#105)
<!--
Thanks for your contribution! Please read this comment in its entirety.
It's quite important.
When a contributor merges the pull request, the title and the
description will be used to build the merge commit!

### Pull Request TITLE

It should be in the following format:

[ 12345: Summary of the changes made ] Where 12345 is the corresponding
Github Issue

OR

[ Summary of the changes made ] If it's solving something trivial, like
fixing a typo.
-->

### Linked issues
<!--
Link the Issue(s) this Pull Request is related to.

Each PR should link to at least one issue, in the form:

Use one line for each Issue. This allows auto-closing the related issue
when the fix is merged.

Fixes #12345
Fixes #54345
-->

Fixes #100 

### Summarize your change.

Fixing MacOS build error in libresample when building with CMake 3.26.

### Describe the reason for the change.

The MacOS build is currently broken due to the faulty project being C
language project and CMake Mac/darwin compiler options are mixed C/C++.

### Describe what you have tested and on which operating system.

Building successfully on MacOS version Ventura 13.2.1.

### Add a list of changes, and note any that might need special
attention during the review.

I have removed standard libc++ flags as they are added by CMake
correctly without adding them to the C compiler.

---------

Signed-off-by: Inouk Babin-Ruel <inouk.babin.ruel@autodesk.com>
Co-authored-by: Kerby Geffrard <kerby.geffrard@autodesk.com>