Fix PySide2 build break with recent macOS/Xcode.app (#110)
<!--
Thanks for your contribution! Please read this comment in its entirety.
It's quite important.
When a contributor merges the pull request, the title and the
description will be used to build the merge commit!

### Pull Request TITLE

It should be in the following format:

[ 12345: Summary of the changes made ] Where 12345 is the corresponding
Github Issue

OR

[ Summary of the changes made ] If it's solving something trivial, like
fixing a typo.
-->

### Summarize your change.

Patch the PySide2 build to not use libXSLT.

### Describe the reason for the change.

Since the last update of MacOS (or Xcode.app?), we can't build PySide2
anymore. The issue seems caused by the way PySide2 finds and uses
libXSLT. It's mixing multiple versions of MacOSX.sdk.

### Describe what you have tested and on which operating system.

MacOS 13.3, Xcode 14.3.

Signed-off-by: Kerby Geffrard <kerby.geffrard@autodesk.com>