Fix readthedocs publication fails (#112)
### Linked issues
None

### Summarize your change.

Fix readthedocs build fails

### Describe the reason for the change.

ReadTheDocs builds are failing on the main Open RV repo. The reason was
the the build would try to run pip as specified in .readthedocs.yaml.
The inclusion of pip was a mistake.

### Describe what you have tested and on which operating system.

Built successfully in OpenRV fork.

### Add a list of changes, and note any that might need special
attention during the review.

Edited: .readthedocs.yaml: removed the pip requirements and its
associated build dir.

---------

Signed-off-by: Martin Chesnay <104032692+mchesnay@users.noreply.github.com>
Signed-off-by: Inouk Babin-Ruel <inouk.babin.ruel@autodesk.com>