Fix building on MacOS >= 13.3 (#113)
### Fix Build on Mac OS >= 13.3

### Summarize your change.
Building on the Mac OS update from last week fails the build. The reason
for this is that the clang headers used in building pyside are too old
and not compatible with the latest Mac SDK. There are pages and pages
full of compilation errors when attempting to compile pyside.

To fix it, we will detect the version of clang on Mac and download the
headers matching that version, even though this is not the version
recommended by PySide, we don't have much choice.

### Describe the reason for the change.
Fixing the build

### Describe what you have tested and on which operating system.
Tested on Mac OS 13.3. and CentOS7.

Signed-off-by: Roger Nelson <roger.nelson@autodesk.com>
Signed-off-by: Inouk Babin-Ruel <inouk.babin.ruel@autodesk.com>