OpenAssetIO + OTIO + OpenRV demo (#114)
### OpenAssetIO + OTIO + OpenRV demo

### Summarize your change.

https://github.com/rogernelson/OpenRV/pull/1

### Describe the reason for the change.

This is a demo in a new branch to collaborate with MovieLabs and
OpenAssetIO. This initial commit adds a new OpenAssetIO package to
OpenRV for resolving URIs and fetching related references. The resolve
happens on the `incoming-source-path` event and a new event is triggered
from the OTIO pre-clip hook to send to the OpenAssetIO package. This
event will fetch related references on the URI in the `target_url`
field. If there are any, the hook will add the returned references to
the existing media_references in OTIO. In turn, these will be resolved
into media when they are imported by the OTIO code (since it will
trigger the `incoming-source-path` event).

The OpenAssetIO package will only initialize if the
`OPENASSETIO_DEFAULT_CONFIG` environment is set which also initializes
the OpenAssetIO manager. So for users not using it, the package will
have no effect.

I've also added some demo-specific files in an example directory for
running the package -- these include config files which are required to
set up the example asset manager (BAL), an OTIO example file with a BAL
URI, and a traits definition file (which will resolve paths based on
environment variables). This allows the demo to be run by supplying the
`OPENASSETIO_DEFAULT_CONFIG` environment variable and `RV_ASSET_{1,2,3}`
to file paths on the local system.

### Describe what you have tested and on which operating system.

Tested on Mac OS 13.3

---------

Signed-off-by: Roger Nelson <roger.nelson@autodesk.com>