Fix Mac and Linux builds (#121)
Fix broken Linux and Mac builds

Fixes https://github.com/AcademySoftwareFoundation/OpenRV/issues/106

The previous update of the OpenEXR library breaks the build since the
version number is part of the library name on Mac and Linux. This branch
just updates the version string the library to the correct version. To
make this easier to do in the future, I just created a variable so it
only needs to be changed in one place.

Tested on Linux.

Signed-off-by: Roger Nelson <roger.nelson@autodesk.com>