Fix intermittent No Audio on the first open/play (#124)
Fix intermittent No Audio on the first open/play [SG-29287+SG-28578]

### Summarize your change.

#### Problem
In some hard to reproduce situations, after loading some media, the
'Audio BUFFERING' HUD indicator was displayed and stuck and the cache
was never filled. Starting a playback would end up with no audio.

#### Cause
In a very specific code path, the audio cache is cleared but the
m_audioCacheComplete flag is NOT reset. This prevented the audio cache
thread from waking up because the IPGraph::maybeDispatchAudioThread()
logic depends on the m_audioCacheComplete flag.

#### Solution
Reset the m_audioCacheComplete flag whenever the audio cache is cleared.
I searched for them all and there was only one place where the audio
cache was cleared and the m_audioCacheComplete was not reset. This is
what I fixed in this commit.

### Describe the reason for the change.

This commit fixes 2 audio related issues:
Fixed an issue where the audio would not play when launching RV through
python commands in a terminal when the Audio Cache option was active
(Scrubbing On By Default (Cache All Audio)). [SG-29287]
Fixed an intermittent issue in RV where no audio would play the first
time RV the application was open or starting playback. [SG-28578]

Note that this fix originates from the RV 2023 commercial release.

(https://community.shotgridsoftware.com/t/rv-2023-0-0-release-is-available/17227)

### Describe what you have tested and on which operating system.
This commit was successfully built on all 3 platforms and tested on
macOS Monterey.

Signed-off-by: Bernard Laberge
[bernard.laberge@autodesk.com](mailto:bernard.laberge@autodesk.com)

Signed-off-by: Bernard Laberge <bernard.laberge@autodesk.com>
Signed-off-by: Inouk Babin-Ruel <inouk.babin.ruel@autodesk.com>