Fix corruption when playing media with inserts (#123)
Fix corruption when playing media with inserts [SG-29243]

### Summarize your change.

##### Problem 
A studio provided us with a ProRes media clip that played back fine in
RV until it reached the middle section where it showed corruption for
the last half of the video.
QuickTime Player on Mac was able to playback the same clip without
corruption.

#### Cause 
Note that RV decodes this ProRes video clip using FFmpeg + AppleProRes
decoding lib.
The issue here is exclusively caused by the FFmpeg wrapper used in RV:
src/lib/image/MovieFFMpeg/MovieFFMpeg.cpp
The video corruption is caused by something we had never had to deal
with before in RV: a change of native video pixel format within the same
video track. In this particular clip, the native video format changes
from yuv422p16le to uyvy422 at about the middle section of the clip.
This broke the original assumption in MovieFFMpeg.cpp that the exact
same format conversion context once allocated, could be reused for all
the frames belonging to the same video track.
This clip provided by the client proved this assumption to be wrong. 

#### Solution 
Two low risk changes were made to address this issue. 

Change1: Now we are always refreshing the track->imgConvertContext via
the sws_getCachedContext() function which is perfectly adapted to our
needs.

Change2: I had to change the order to seek to the righ position in the
stream BEFORE looking at the videoCodecContext->pix_fmt to determine the
format conversion because the videoCodecContext->pix_fmt gets updated
only after findImageWithBestTimestamp() is executed.
So even though many lines seemed to be changed, only the order was
changed.

### Describe the reason for the change.

This commit fixes an issue where the Player would display corrupted
images when playing back media with inserts. [SG-29243]

Note that this fix originates from the RV 2023 commercial release.


(https://community.shotgridsoftware.com/t/rv-2023-0-0-release-is-available/17227)

### Describe what you have tested and on which operating system.

This commit was successfully built on all 3 platforms and tested on
macOS Monterey.

Signed-off-by: Bernard Laberge <bernard.laberge@autodesk.com>

Signed-off-by: Bernard Laberge <bernard.laberge@autodesk.com>