Rename the MMR switch groups to something more meaningful (#128)
Rename the MMR switch groups to something more meaningful [SG-28740]

### Summarize your change.

#### Problem:
With the new Multiple Media Representations feature, the ui names of the
source nodes and switch nodes were not really meaningful.

#### Solution:
In this commit, the media representation name is used as the ui name by
default.

### Describe the reason for the change.

The MMR switch groups and sources displayed in the Session Manager now
use names that are now more meaningful. [SG-28740]

Note that this fix originates from the RV 2023 commercial release.

(https://community.shotgridsoftware.com/t/rv-2023-0-0-release-is-available/17227)

### Describe what you have tested and on which operating system.

This commit was successfully built on all 3 platforms and tested on
macOS Monterey.

Signed-off-by: Bernard Laberge <bernard.laberge@autodesk.com>

Signed-off-by: Bernard Laberge <bernard.laberge@autodesk.com>
Signed-off-by: Inouk Babin-Ruel <inouk.babin.ruel@autodesk.com>