Fix viewer not updated after RV reload() command (#130)
Fix viewer not updated after RV reload() command [SG-30348]

### Summarize your change.

#### Problem 
The RV reload() command which is used by the "RV/Tools/Force Reload..."
menu options instructs RV to reload media files.
The command works as expected except for the current frame which never
seems to be updated.
However, if one selects the next frame and then goes back to the
previous frame then the viewer gets updated.
This clunky work around is not available when viewing a single frame
which is why the client is interested into a fix for this issue.
The client tested the issue on multiple versions of RV and confirms that
this regression was introduced in RV 7.1 (2015)

#### Cause 
I found the cause of the regression which had to do with the RV's Image
Renderer refactoring that took place in RV 2015.
In the refactored image renderer, there is a simple cache of uploaded
textures for the current frame (ImageRenderer::m_uploadedTextures).
Since that when reloading a media file, the unique identifier associated
with the frame buffer does not change (because it is the exact same
filename that gets reloaded).

#### Solution 
Added a ImageRenderer::freeUploadedTextures() that takes care of freeing
the uploaded textures cache and use it in the reload() RV command. This
means that this fix's impact is limited to the RV reload() command which
greatly minimizes the risks.

Note that I also removed the clearImages() method which was not defined
(nor used thankfully)

### Describe the reason for the change.

This commit fixes an issue about the Player not updating the current
frame when an RV reload() command or Force Reload menu option is being
done. [SG-30348]

Note that this fix originates from the RV 2023 commercial release.


(https://community.shotgridsoftware.com/t/rv-2023-0-0-release-is-available/17227)

### Describe what you have tested and on which operating system.

This commit was successfully built on all 3 platforms and tested on
macOS Monterey.

Signed-off-by: Bernard Laberge <bernard.laberge@autodesk.com>

Signed-off-by: Bernard Laberge <bernard.laberge@autodesk.com>