OTIO Import: Fix blend order and global bounds (#139)
### OTIO Import: Fix blend order and global bounds

This branch fixes two issues and is brought over from the commercial RV
branch:
- Global image bounds on import
- Blend order

Previously, we would just use the first clip's image bounds to set the
global bounds (ie canvas size) which works fine if all clips have the
same bounds. However, if not, anything outside of the first clip's
bounds will be clipped. To fix this, we'll simply calculate the union of
all bounds and use that as the canvas size. The canvas will initially
still be framed around the first clip, but with this fix zooming out
ensures the image outside of those bounds will be displayed.

The blend order in a stack, for some reason (I assume legacy), is not
actually a stack. The first input is displayed on top instead of the
last. To fix this, set the `supportReversedOrderBlending` property so
that the last input is on top. This will ensure for simultaneously
display clips (an OTIO stack) -- like a picture-in-picture effect done
with bounds for example -- the last clip will be drawn on top of the
first.

There also a few minor style changes related to running `black` on the
file.

Tested on Mac and Linux.

Signed-off-by: Roger Nelson <roger.nelson@autodesk.com>