Fix Mac install (#142)
This branch fixes the Mac installation.

The problem is the CMake variable `PROJECT_SOURCE_DIR` is not available
on the install step (only configure and build). Since no variables
involving the source dir are present at installation time, the only
suitable replacement I could find was to use a path relative to the
current directory `CMAKE_CURRENT_LIST_DIR`.

Tested on Mac 13.3.1

Signed-off-by: Roger Nelson <roger.nelson@autodesk.com>
Signed-off-by: Inouk Babin-Ruel <inouk.babin.ruel@autodesk.com>