Add a section on how to delete sources via the api (#153)
### Linked issues

None. This is a trivial fix.

### Summarize your change.

Describe how to delete cleanly a source from the image graph. 

### Describe the reason for the change.

The documentation describes how to use `addSourcexxx()` to add a source
to the image graph, but did not explain that simply deleting the
returned node did not clean up the graph. This update fixes that.

### Add a list of changes, and note any that might need special
attention during the review.

Only updated file is [Ref Manual: Chapter
Two](docs/rv-manuals/rv-reference-manual/rv-reference-manual-chapter-two.md)
Note that some trivial markdown clean up also happened. The only area of
concern is section 2.3.2 Deleting a Source.

Signed-off-by: Martin Chesnay <Martin.Chesnay@autodesk.com>