Strip debug symbols from binaries when installing release on Linux (#154)
[Strip debug symbols from binaries when installing release on Linux]

### Summarize your change.

This brach adds some additional code to the RV installation step. When
executed it will check the file to be installed to see if its a binary,
and if so strip out any debugging symbols it finds. Certain mime types
that cause errors with the strip command have been filtered out, and an
error message will be displayed on any types that may fail in the future
indicating that they may need to be added to the ignore list.

I did this at installation time instead of build time (with the -s
option in CXX_FLAGS) so that we would also strip third-party binaries
that we install, since we don't pass out CXX_FLAGS down to them when we
build them.

### Describe the reason for the change.

It will reduce the size of the linux installation and should improve
load performance.

### Describe what you have tested and on which operating system.

Built on Linux Rocky 9.  Reduced installation size from 3.5 to 1.3 GB.

Signed-off-by: Roger Nelson <roger.nelson@autodesk.com>